requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
flask-redis==0.4.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
python-dotenv==0.10.3
redis==3.3.8
requests==2.22.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
